requirements.txt
mcp[cli]
requests
pydantic
python-dotenv
fastapi
uvicorn